Fixed voting failure issue

+++ contracts/RocketNodeValidator.sol
@@ -15,7 +15,7 @@ contract RocketNodeValidator is RocketBase {
     /** Events */
 
     event NodeVoteCast (
-        uint128 epoch,
+        uint256 epoch,
         address minipool_address,
         bytes vote_message
     );
@@ -39,7 +39,7 @@ contract RocketNodeValidator is RocketBase {
     /// @param _epoch The epoch number voting relates to
     /// @param _minipool_address The address of the minipool that should cast the votes
     /// @param _vote_message Vote message to be sent to Casper
-    function nodeVote(uint128 _epoch, address _minipool_address, bytes _vote_message) public onlyRegisteredNode(msg.sender) returns(bool) {
+    function nodeVote(uint256 _epoch, address _minipool_address, bytes _vote_message) public onlyRegisteredNode(msg.sender) returns(bool) {
         RocketPoolInterface rocketPool = RocketPoolInterface(rocketStorage.getAddress(keccak256("contract.name", "rocketPool")));
         require(rocketPool.vote(_epoch, _minipool_address, _vote_message));
         NodeVoteCast(_epoch, _minipool_address, _vote_message);
